Moving execute_line, and execute_block into Lexicon
